+++ News.md
@@ -3,6 +3,8 @@
 - `TreeSitter::Node` is enumerable.
 - `TreeSitter::{QueryCaptures, QueryMatches}` are enumerable.
 - `TreeStand::Node` now supports query predicates from `TreeSitter`.
+- `TreeSitter::QueryMatches` now has a `each_capture_hash` method returning an `Enumerator<Hash<String, Node>>`,
+  the rough equivalent of what `TreeStand::Node#query` returns.
 
 # v1.3.0